+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Custom logger decorators can be registered on the Factory and applied via config
+- New `Decorator\DefaultContext` which allows default context values to be set for all logs.
+  Use directly or add to config, eg. `'defaultContext' => ['foo'=>'bar']`
 
 ### Changed
 - Renamed `LevelFilter` to `Decorator\LevelFilter`